+++ desktop/cmp/input/CodeInput.js
@@ -5,13 +5,15 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 import {HoistInput} from '@xh/hoist/cmp/input';
-import {box, fragment, hbox} from '@xh/hoist/cmp/layout';
+import {box, filler, fragment, frame, hbox, label, span} from '@xh/hoist/cmp/layout';
 import {elemFactory, HoistComponent, LayoutSupport, XH} from '@xh/hoist/core';
 import {button} from '@xh/hoist/desktop/cmp/button';
 import {clipboardButton} from '@xh/hoist/desktop/cmp/clipboard';
+import {textInput} from '@xh/hoist/desktop/cmp/input/TextInput';
+import {toolbar, toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
 import {Icon} from '@xh/hoist/icon';
 import {dialog, textArea} from '@xh/hoist/kit/blueprint';
-import {bindable} from '@xh/hoist/mobx';
+import {action, bindable, observable} from '@xh/hoist/mobx';
 import {withDefault} from '@xh/hoist/utils/js';
 import * as codemirror from 'codemirror';
 import 'codemirror/addon/fold/brace-fold.js';
@@ -22,9 +24,11 @@ import 'codemirror/addon/lint/lint.css';
 import 'codemirror/addon/lint/lint.js';
 import 'codemirror/addon/scroll/simplescrollbars.css';
 import 'codemirror/addon/scroll/simplescrollbars.js';
+import 'codemirror/addon/search/searchcursor.js';
+import 'codemirror/addon/selection/mark-selection.js';
 import 'codemirror/lib/codemirror.css';
 import 'codemirror/theme/dracula.css';
-import {defaultsDeep, isFunction} from 'lodash';
+import {compact, defaultsDeep, isEmpty, isEqual, isFunction} from 'lodash';
 import PT from 'prop-types';
 import ReactDOM from 'react-dom';
 import './CodeInput.scss';
@@ -43,12 +47,6 @@ import './CodeInput.scss';
 @LayoutSupport
 export class CodeInput extends HoistInput {
 
-    /** @member {CodeMirror} - a CodeMirror editor instance. */
-    editor;
-
-    @bindable fullScreen = false;
-    baseClassName = 'xh-code-input';
-
     static propTypes = {
         ...HoistInput.propTypes,
 
@@ -65,6 +63,13 @@ export class CodeInput extends HoistInput {
         editorProps: PT.object,
 
         /**
+         * True to enable case-insensitive searching within the input. Default false, except in
+         * fullscreen mode, where search will be shown unless explicitly *disabled*. Note that
+         * enabling search forces the display of a toolbar, regardless of `showToolbar` prop.
+         */
+        enableSearch: PT.bool,
+
+        /**
          * Callback to autoformat the code. Given the unformatted code, this should return a
          * properly-formatted copy.
          */
@@ -74,8 +79,8 @@ export class CodeInput extends HoistInput {
         linter: PT.func,
 
         /**
-         * A CodeMirror language mode - default none (plain-text).
-         * See the CodeMirror docs ({@link https://codemirror.net/mode/}) regarding available modes.
+         * A CodeMirror language mode - default none (plain-text). See the CodeMirror docs
+         * ({@link https://codemirror.net/mode/}) regarding available modes.
          * Applications must import any mode they wish to enable.
          */
         mode: PT.string,
@@ -95,22 +100,81 @@ export class CodeInput extends HoistInput {
          */
         showFormatButton: PT.bool,
 
-        /** True (default) to display Fullscreen button at bottom-right of input. */
-        showFullscreenButton: PT.bool
+        /** True (default) to display fullscreen button at bottom-right of input. */
+        showFullscreenButton: PT.bool,
+
+        /**
+         * True to display action buttons and/or find functionality in a dedicated bottom toolbar.
+         * Default is false unless enableSearch==true or in fullscreen mode. When false, enabled
+         * action buttons show only when the input focused and float in the bottom-right corner.
+         */
+        showToolbar: PT.bool
     };
 
+
+    /** @member {CodeMirror} - a CodeMirror editor instance. */
+    editor;
+
+    // Support for internal search feature.
+    cursor = null;
+    @bindable query = '';
+    @observable currentMatchIdx = -1;
+    @observable.ref matches = [];
+    get matchCount() {return this.matches.length}
+
+    @observable fullScreen = false;
+    baseClassName = 'xh-code-input';
+
     get commitOnChange() {return withDefault(this.props.commitOnChange, true)}
+
     get showCopyButton() {return withDefault(this.props.showCopyButton, false)}
+
     get showFullscreenButton() {return withDefault(this.props.showFullscreenButton, true)}
 
     get showFormatButton() {
         const {disabled, readonly, formatter, showFormatButton} = this.props;
-        return (
-            !disabled &&
-            !readonly &&
-            isFunction(formatter) &&
-            withDefault(showFormatButton, true)
-        );
+        return (!disabled && !readonly && isFunction(formatter) && withDefault(showFormatButton, true));
+    }
+
+    get showAnyActionButtons() {
+        const {showCopyButton, showFormatButton, showFullscreenButton} = this;
+        return showCopyButton || showFormatButton || showFullscreenButton;
+    }
+
+    get showSearchInput() {
+        return withDefault(this.props.enableSearch, this.fullScreen);
+    }
+
+    get showToolbar() {
+        const {props, showSearchInput, showAnyActionButtons, fullScreen} = this;
+        // Always show if showing searchInput - it's the only place searchInput can live.
+        if (showSearchInput) return true;
+        // Show if prop enabled and at least one action button.
+        if (props.showToolbar && showAnyActionButtons) return true;
+        // Show if fullscreen w/buttons and prop not explicitly *disabled*.
+        return (fullScreen && showAnyActionButtons && props.showToolbar !== false);
+    }
+
+    get actionButtons() {
+        const {showCopyButton, showFormatButton, showFullscreenButton, editor} = this;
+        return compact([
+            showCopyButton ? clipboardButton({
+                text: null,
+                title: 'Copy to clipboard',
+                successMessage: 'Contents copied to clipboard',
+                getCopyText: () => editor.getValue()
+            }) : null,
+            showFormatButton ? button({
+                icon: Icon.magic(),
+                title: 'Auto-format',
+                onClick: () => this.onAutoFormat()
+            }) : null,
+            showFullscreenButton ? button({
+                icon: this.fullScreen ? Icon.collapse() : Icon.expand(),
+                title: this.fullScreen ? 'Exit full screen' : 'Full screen',
+                onClick: () => this.toggleFullScreen()
+            }) : null
+        ]);
     }
 
     constructor(props, context) {
@@ -142,6 +206,18 @@ export class CodeInput extends HoistInput {
                 this.editor.setOption('readOnly',  disabled || readonly);
             }
         });
+
+        this.addReaction({
+            track: () => this.query,
+            run: (query) => {
+                if (query?.trim()) {
+                    this.findAll();
+                } else {
+                    this.clearSearchResults();
+                }
+            },
+            debounce: 300
+        });
     }
 
     render() {
@@ -162,7 +238,7 @@ export class CodeInput extends HoistInput {
                 isOpen: true,
                 canOutsideClickClose: true,
                 item: this.renderInput({flex: 1}),
-                onClose: () => this.setFullScreen(false)
+                onClose: () => this.toggleFullScreen()
             }),
             box({
                 className: 'xh-code-input--placeholder',
@@ -172,6 +248,7 @@ export class CodeInput extends HoistInput {
     }
 
     renderInput(props) {
+        const {showToolbar} = this;
         return box({
             items: [
                 textArea({
@@ -179,46 +256,88 @@ export class CodeInput extends HoistInput {
                     ref: this.manageCodeEditor,
                     onChange: this.onChange
                 }),
-                this.renderActionButtons()
+                showToolbar ? this.renderToolbar() : this.renderActionButtons()
             ],
-
             className: this.getClassName(),
             onBlur: this.onBlur,
             onFocus: this.onFocus,
-
             ...props
         });
     }
 
-    renderActionButtons() {
-        const {showCopyButton, showFormatButton, showFullscreenButton} = this;
+    renderToolbar() {
+        const {actionButtons} = this,
+            searchInput = this.renderSearchInput();
 
-        if (!this.hasFocus || (!showCopyButton && !showFormatButton && !showFullscreenButton)) {
-            return null;
-        }
-
-        const {fullScreen, editor} = this;
-        return hbox({
-            className: 'xh-code-input__action-buttons',
+        return toolbar({
+            className: 'xh-code-input__toolbar',
             items: [
-                showCopyButton ? clipboardButton({
-                    text: null,
-                    title: 'Copy to clipboard',
-                    successMessage: 'Contents copied to clipboard',
-                    getCopyText: () => editor.getValue()
-                }) : null,
-                showFormatButton ? button({
-                    icon: Icon.magic(),
-                    title: 'Auto-format',
-                    onClick: () => this.onAutoFormat()
-                }) : null,
-                showFullscreenButton ? button({
-                    icon: fullScreen ? Icon.collapse() : Icon.expand(),
-                    title: fullScreen ? 'Exit full screen' : 'Full screen',
-                    onClick: () => this.setFullScreen(!fullScreen)
-                }) : null
+                filler(),
+                searchInput,
+                toolbarSep({omit: !searchInput || isEmpty(actionButtons)}),
+                ...actionButtons
             ]
         });
+    }
+
+    renderSearchInput() {
+        const {showSearchInput, cursor, currentMatchIdx, matchCount} = this;
+        if (!showSearchInput) return null;
+
+        return fragment(
+            // Frame wrapper added due to issues with textInput not supporting all layout props as it should.
+            frame({
+                flex: 1,
+                maxWidth: 400,
+                item: textInput({
+                    width: null,
+                    flex: 1,
+                    model: this,
+                    bind: 'query',
+                    leftIcon: Icon.search(),
+                    enableClear: true,
+                    commitOnChange: true,
+                    onKeyDown: (e) => {
+                        if (e.key != 'Enter') return;
+                        if (!cursor) {
+                            this.findAll();
+                        } else if (e.shiftKey) {
+                            this.findPrevious();
+                        } else {
+                            this.findNext();
+                        }
+                    }
+                })
+            }),
+            label({
+                className: 'xh-code-input__label',
+                item: matchCount ?
+                    `${currentMatchIdx + 1} / ${matchCount}` :
+                    span({item: '0 results', className: 'xh-text-color-muted'})
+            }),
+            button({
+                icon: Icon.arrowUp(),
+                title: 'Find previous (shift+enter)',
+                disabled: !matchCount,
+                onClick: () => this.findPrevious()
+            }),
+            button({
+                icon: Icon.arrowDown(),
+                title: 'Find next (enter)',
+                disabled: !matchCount,
+                onClick: () => this.findNext()
+            })
+        );
+    }
+
+    renderActionButtons() {
+        const {hasFocus, actionButtons} = this;
+
+        return (hasFocus && actionButtons.length) ?
+            hbox({
+                className: 'xh-code-input__action-buttons',
+                items: actionButtons
+            }) : null;
     }
 
     //------------------
@@ -227,6 +346,7 @@ export class CodeInput extends HoistInput {
     manageCodeEditor = (textAreaComp) => {
         if (textAreaComp) {
             this.editor = this.createCodeEditor(textAreaComp);
+            this.preserveSearchResults();
         }
     };
 
@@ -272,7 +392,7 @@ export class CodeInput extends HoistInput {
             readOnly: disabled || readonly,
             gutters,
             lint: linter ? {getAnnotations: linter} : false,
-            autofocus: autoFocus
+            autoFocus
         };
     }
 
@@ -282,6 +402,7 @@ export class CodeInput extends HoistInput {
 
     handleEditorChange = (editor) => {
         this.noteValueChange(editor.getValue());
+        if (this.cursor) this.clearSearchResults();
     };
 
     onAutoFormat = () => {
@@ -300,6 +421,99 @@ export class CodeInput extends HoistInput {
         }
     }
 
+    @action
+    toggleFullScreen() {
+        this.fullScreen = !this.fullScreen;
+    }
+
+    //------------------------
+    // Local Searching
+    //------------------------
+    @action
+    findAll() {
+        this.clearSearchResults();
+        if (!this.query?.trim()) return;
+
+        this.cursor = this.editor.getSearchCursor(this.query, 0, true);
+
+        const {cursor, editor} = this,
+            newMatches = [];
+
+        while (cursor.findNext()) {
+            const anchor = cursor.from(),
+                head = cursor.to();
+            newMatches.push({
+                anchor,
+                head,
+                textMarker: editor.markText(anchor, head, {className: 'xh-code-input--highlight'})
+            });
+        }
+
+        this.matches = newMatches;
+        if (newMatches.length) {
+            this.findNext();
+        } else {
+            this.currentMatchIdx = -1;
+        }
+    }
+
+    @action
+    findNext() {
+        const {editor, query, cursor, matchCount} = this;
+        if (!cursor || !matchCount) return;
+
+        if (cursor.findNext(query)) {
+            this.handleCursorMatchUpdate();
+        } else {
+            // Loop around
+            this.cursor = editor.getSearchCursor(query, 0, true);
+            this.findNext();
+        }
+    }
+
+    @action
+    findPrevious() {
+        const {editor, query, cursor, matches, matchCount} = this;
+        if (!cursor || !matchCount) return;
+
+        if (cursor.findPrevious(query)) {
+            this.handleCursorMatchUpdate();
+        } else {
+            // Loop around
+            this.cursor = editor.getSearchCursor(query, matches[matchCount - 1].head, true);
+            this.findPrevious();
+        }
+    }
+
+    @action
+    handleCursorMatchUpdate() {
+        const {editor, cursor, matches} = this,
+            from = cursor.from(),
+            to = cursor.to();
+        editor.scrollIntoView({from, to}, 50);
+        editor.setSelection(from, to);
+        this.currentMatchIdx = matches.findIndex(match => isEqual(match.anchor, from));
+    }
+
+    preserveSearchResults() {
+        const {matches, editor} = this;
+        matches.forEach(match => {
+            match.textMarker = editor.markText(match.anchor, match.head, {className: 'xh-code-input--highlight'});
+        });
+    }
+
+    @action
+    clearSearchResults() {
+        this.cursor = null;
+        this.currentMatchIdx = -1;
+        this.matches.forEach(match => match.textMarker.clear());
+        this.matches = [];
+    }
+
+
+    //------------------------
+    // Other Lifecycle
+    //------------------------
     destroy() {
         // Cleanup editor component as per CodeMirror docs.
         if (this.editor) this.editor.toTextArea();